+++ plugins/zulip/zulipbot/config/resources/db_init_data.yaml
