+++ plugins/zulip/zulipbot/config/resources/db_structure.yaml
